pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "torch>=2.4.1",
    "flux>=1.3.5",
    "pillow>=10.4.0",
    "flask-caching>=2.3.0",
    "flask-limiter>=3.8.0",
    "openai>=1.51.0",
    "requests>=2.32.3",
]
